refactor: Mukjipa의 decideWinner 함수 분리, enum 파일 분리, 매직리터럴 제거
