+++ RockPaperScissors/RockPaperScissors/Decision.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum Decision {
+    static let winNumber = [1, -2]
+    static let defeatNumber = [2, -1]
+}
