+++ RockPaperScissors/RockPaperScissors/ErrorMessage.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum ErrorMessage {
+    static let wrongInput = "잘못된 입력입니다. 다시 시도해주세요."
+}
