+++ RockPaperScissors/RockPaperScissors/Player.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum Player {
+    static let user = "사용자"
+    static let computer = "컴퓨터"
+}
