+++ RockPaperScissors/RockPaperScissors/PunctuationMarks.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum PunctuationMarks {
+    static let colon = " : "
+    static let emptyString = ""
+}
