+++ RockPaperScissors/RockPaperScissors/StartAndEndMessage.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum StartAndEndMessage {
+    static let startOfFirstGame = "가위(1), 바위(2), 보(3)! <종료 : 0>"
+    static let startOfSecondGame = "묵(1), 찌(2), 빠(3)! <종료 : 0>"
+    static let endOfGame = "게임 종료"
+}
